requirements.txt
streamlit>=1.25
faiss-cpu>=1.7.4
google-genai>=1.0.0
langchain-google-genai>=2.1.7
langchain>=0.3.20
langchain-huggingface>=0.3.0
sentence-transformers>=2.2.2
duckduckgo-search>=2.0.0
numpy>=1.24
google-auth>=2.20.0
